refactor(validation): split src/lib/validation.ts into per-source modules
